requirements.txt
langchain>=1.1.0
langchain-openai>=0.1.22
langchain-anthropic>=1.2.0
langchain-community>=0.2.9
langchain-text-splitters>=0.2.2
langgraph>=1.0.2
deepagents>=0.3.0
faiss-cpu>=1.8.0
rank_bm25>=0.2.2
python-dotenv>=1.0.1
requests>=2.32.0
mcp>=1.1.0
fastapi>=0.110.0
uvicorn>=0.30.0
httpx>=0.27.0
anyio>=4.0.0
